--- Login-bttn.module.css
@@ -1,163 +1,0 @@
-.Login-div {
-    text-align: right;
-    justify-content: right;
-    right: 0%;
-    top: 2.5%;
-    width: 1000px;
-    height: 500px;
-    float: right;
-    position: absolute;
-    display: flex;
-}
-
-.Login-bttn {
-    margin: 5px;
-    padding: 2px;
-    width: 80px;
-    height: 35px;
-    border: 4px solid #000;
-    color: #000;
-    border-radius: 20px;
-    transition: all 0.5s;
-    background: #fff;
-    font-family: 'Black Han Sans', sans-serif;
-    cursor: pointer;
-    box-shadow: inset 0 0 0 0 #ffffff;
-}
-
-    .Login-bttn:hover {
-        box-shadow: inset 0 0 0 100vmax #000000;
-        color: #fff;
-        transform: none;
-    }
-
-    .Login-bttn.active {
-        margin: 5px;
-        padding: 2px;
-        width: 80px;
-        height: 35px;
-        border: 4px solid #000;
-        color: #fff;
-        border-radius: 20px;
-        transition: all 0.5s;
-        background: #000;
-        font-family: 'Black Han Sans', sans-serif;
-        cursor: pointer;
-        box-shadow: inset 0 0 0 0 #000000;
-    }
-
-    .Login-bttn.close {
-        visibility: hidden;
-    }
-
-.Login-popup {
-    text-align: center;
-    margin: 10px;
-    padding: 10px;
-    float: left;
-    top: 15%;
-    right: 40%;
-    position: absolute;
-    width: 500px;
-    height: 180px;
-    border: 4px solid black;
-    font-family: 'Black Han Sans', sans-serif;
-    background-color: #fff;
-    z-index: 1;
-    display: block;
-    visibility: hidden;
-}
-
-    .Login-popup.active {
-        animation: scale-in-center 0.5s cubic-bezier(0.175, 0.885, 0.320, 1.275);
-        visibility: visible;
-    }
-
-    .Login-popup.close {
-        animation: scale-out-center 0.5s cubic-bezier(0.175, 0.885, 0.320, 1.275);
-        transition: visibility 0.5s linear;
-    }
-
-.Popuptext-div {
-    width: auto;
-    height: 50%;
-    border: 4px solid black;
-    padding: 20px;
-    font-family: 'Black Han Sans', sans-serif;
-    background-color: #fff;
-    text-align: center;
-    margin: 5px;
-    white-space: pre-wrap;
-}
-
-.Popupbttn-div {
-    width: 100px;
-    height: 30px;
-    position: absolute;
-    margin-top: 10px;
-    left: 80%;
-}
-
-@keyframes File-popup {
-    0% {
-        -webkit-transform: rotateX(-100deg);
-        transform: rotateX(-100deg);
-        -webkit-transform-origin: top;
-        transform-origin: top;
-        opacity: 0;
-    }
-
-    100% {
-        -webkit-transform: rotateX(0deg);
-        transform: rotateX(0deg);
-        -webkit-transform-origin: top;
-        transform-origin: top;
-        opacity: 1;
-    }
-}
-
-@keyframes File-hide {
-    0% {
-        -webkit-transform: rotateX(0deg);
-        transform: rotateX(0deg);
-        -webkit-transform-origin: top;
-        transform-origin: top;
-        opacity: 1;
-    }
-
-    100% {
-        -webkit-transform: rotateX(-100deg);
-        transform: rotateX(-100deg);
-        -webkit-transform-origin: top;
-        transform-origin: top;
-        opacity: 0;
-    }
-}
-
-@keyframes scale-in-center {
-    0% {
-        -webkit-transform: scale(0);
-        transform: scale(0);
-        opacity: 1;
-    }
-
-    100% {
-        -webkit-transform: scale(1);
-        transform: scale(1);
-        opacity: 1;
-    }
-}
-
-@keyframes scale-out-center {
-    0% {
-        -webkit-transform: scale(1);
-        transform: scale(1);
-        opacity: 1;
-    }
-
-    100% {
-        -webkit-transform: scale(0);
-        transform: scale(0);
-        opacity: 1;
-    }
-}
