+++ client/src/api/firebase.js
@@ -1,7 +1,7 @@
 // Import the functions you need from the SDKs you need
 import { initializeApp } from "firebase/app";
-import { GoogleAuthProvider, getAuth, onAuthStateChanged, signInWithPopup } from "firebase/auth";
-import { useState } from "react";
+import { GoogleAuthProvider, getAuth, signInWithPopup,signOut,onAuthStateChanged } from "firebase/auth";
+
 
 const firebaseConfig = {
   apiKey: process.env.REACT_APP_FIREBASE_APIKEY,
@@ -22,6 +22,14 @@ export async function googleLogin() {
 }
 
 
+export async function googleLogout() {
+  return signOut(auth)
+  .then(() => null)
+}
 
 
-
+export function googleUserChange(callback) {
+  onAuthStateChanged(auth, (user) => {
+    callback(user.email);
+  });
+}
